+++ src/codec/tags/layerStyles/DropShadowStyle.cpp
@@ -21,7 +21,8 @@
 namespace pag {
 std::unique_ptr<BlockConfig> DropShadowStyleTag(DropShadowStyle* style) {
   auto tagConfig = new BlockConfig(TagCode::DropShadowStyle);
-  AddAttribute(tagConfig, &style->blendMode, AttributeType::DiscreteProperty, BlendMode::Normal);
+  AddAttribute(tagConfig, &style->blendMode, AttributeType::DiscreteProperty,
+               static_cast<uint8_t>(BlendMode::Normal));
   AddAttribute(tagConfig, &style->color, AttributeType::SimpleProperty, Black);
   AddAttribute(tagConfig, &style->opacity, AttributeType::SimpleProperty, (Opacity)191);
   AddAttribute(tagConfig, &style->angle, AttributeType::SimpleProperty, 120.0f);
